fix(analysis): unintended queryParams reset when editing reports
We serialize the ordering as null if the
value equals the defaultValue. For some reason
if the serialized value is not null we had some
weird query param behavior where the ordering
param was set to the defaultValue when editing
a report and due to that all other queryParams
where reset.