test(activities): lax current url check
Restarting a task from the past will present the current date in the URL
even though it should not be displayed, since it's the default QP for day.
The QP handling is cumbersome so I decided it's ok to have it there for
this special case. Please improve if possible.