fix(statistics): refactor multiqs to use filtering aggregates
Filtering aggregates are much easier to manage, and can be constructed
using way simpler queries. This improves performance significantly.

The resulting query is now not a multi-level-nested-subquery that needs
to be joined in awkward ways and causing headaches to the PostgreSQL
query planner and optimizer.

Multi-minute queries now take a fraction of a second.