chore(frontend): rewrite report ability to work with ember-data 5.3
this changes `ability.canEdit` into 2 parts, `ability.canEdit` which
checks if for if you are superuser, if the report hasn't been verified
and if the report is your own.

the other parts of the ability, checking for reviewer and supervisee have
been moved into `ability.canAedit`, which is an async ability.

places where this is used first check with `canEdit`, if thats false
they `await canAedit(report)`.

this is ugly and hacky but it works.