fix(statistics): use subquery instead of join to avoid cartesian product
The way we filter in the statistics view, any added filter (that affects
reports) adds a "dimension" to the cartesian product, exploding the total
number of hours reported.

Instead of using JOIN, we do EXISTS(SUBQUERY) now, which should avoid this
issue. Might be a tiny bit slower, but let's try to make it correct first, then fast.

Also fix the catch prefix on the Task statistics filterset - that was wrong
and caused a few incorrect filterings as well.

And also: Add a test to validate multiple-filter behaviour in task stats.