refactor(statistics): introduce a statistics base model based on SQL VIEW
This is still WIP, but I think this is the right way forward.

TODO:

* Further refactor the views, as the customer,project,task viewsets should
  return all the selected objects, regardless of reported sum. I think we
  will need to base those viewsets off of their respective models and
  join/prefetch the report statistic onto it
* Finalize the rest of the refactoring to make it work again

Optionally (and I hope it's not necessary): Build a statistic (database) view
for all statistics endpoints. This would allow us to optimize the whole thing
much more (and improve performance AND readability) at the cost of quite some
more boilerplate code